Split `AssertionScope` specs in multiple files
